feat(cls): [123708596] add new resource (#3335)
* add

* add

* add

* add

+++ tencentcloud/services/cls/resource_tc_cls_cloud_product_log_task_v2.go
@@ -15,13 +15,12 @@ import (
 	"github.com/tencentcloudstack/terraform-provider-tencentcloud/tencentcloud/internal/helper"
 )
 
-func ResourceTencentCloudClsCloudProductLogTask() *schema.Resource {
+func ResourceTencentCloudClsCloudProductLogTaskV2() *schema.Resource {
 	return &schema.Resource{
-		DeprecationMessage: "This resource has been deprecated in Terraform TencentCloud provider version 1.81.188. Please use `tencentcloud_cls_cloud_product_log_task_v2` instead.",
-		Create:             resourceTencentCloudClsCloudProductLogTaskCreate,
-		Read:               resourceTencentCloudClsCloudProductLogTaskRead,
-		Update:             resourceTencentCloudClsCloudProductLogTaskUpdate,
-		Delete:             resourceTencentCloudClsCloudProductLogTaskDelete,
+		Create: resourceTencentCloudClsCloudProductLogTaskV2Create,
+		Read:   resourceTencentCloudClsCloudProductLogTaskV2Read,
+		Update: resourceTencentCloudClsCloudProductLogTaskV2Update,
+		Delete: resourceTencentCloudClsCloudProductLogTaskV2Delete,
 		Importer: &schema.ResourceImporter{
 			State: schema.ImportStatePassthrough,
 		},
@@ -29,24 +28,28 @@ func ResourceTencentCloudClsCloudProductLogTask() *schema.Resource {
 			"instance_id": {
 				Type:        schema.TypeString,
 				Required:    true,
+				ForceNew:    true,
 				Description: "Instance ID.",
 			},
 
 			"assumer_name": {
 				Type:        schema.TypeString,
 				Required:    true,
+				ForceNew:    true,
 				Description: "Cloud product identification, Values: CDS, CWP, CDB, TDSQL-C, MongoDB, TDStore, DCDB, MariaDB, PostgreSQL, BH, APIS.",
 			},
 
 			"log_type": {
 				Type:        schema.TypeString,
 				Required:    true,
+				ForceNew:    true,
 				Description: "Log type, Values: CDS-AUDIT, CDS-RISK, CDB-AUDIT, TDSQL-C-AUDIT, MongoDB-AUDIT, MongoDB-SlowLog, MongoDB-ErrorLog, TDMYSQL-SLOW, DCDB-AUDIT, DCDB-SLOW, DCDB-ERROR, MariaDB-AUDIT, MariaDB-SLOW, MariaDB-ERROR, PostgreSQL-SLOW, PostgreSQL-ERROR, PostgreSQL-AUDIT, BH-FILELOG, BH-COMMANDLOG, APIS-ACCESS.",
 			},
 
 			"cloud_product_region": {
 				Type:        schema.TypeString,
 				Required:    true,
+				ForceNew:    true,
 				Description: "Cloud product region. There are differences in the input format of different log types in different regions. Please refer to the following example:\n- CDS(all log type): ap-guangzhou\n- CDB-AUDIT: gz\n- TDSQL-C-AUDIT: gz\n- MongoDB-AUDIT: gz\n- MongoDB-SlowLog: ap-guangzhou\n- MongoDB-ErrorLog: ap-guangzhou\n- TDMYSQL-SLOW: gz\n- DCDB(all log type): gz\n- MariaDB(all log type): gz\n- PostgreSQL(all log type): gz\n- BH(all log type): overseas-polaris(Domestic sites overseas)/fsi-polaris(Domestic sites finance)/general-polaris(Domestic sites)/intl-sg-prod(International sites)\n- APIS(all log type): gz.",
 			},
 
@@ -59,25 +62,17 @@ func ResourceTencentCloudClsCloudProductLogTask() *schema.Resource {
 			"logset_name": {
 				Type:        schema.TypeString,
 				Optional:    true,
-				Description: "Log set name, it will be automatically created.",
-			},
-
-			"logset_id": {
-				Type:        schema.TypeString,
+				ForceNew:    true,
 				Computed:    true,
-				Description: "Log set ID.",
+				Description: "Log set name, required if `logset_id` is not filled in. If the log set does not exist, it will be automatically created.",
 			},
 
 			"topic_name": {
 				Type:        schema.TypeString,
 				Optional:    true,
-				Description: "The name of the log topic, it will be automatically created.",
-			},
-
-			"topic_id": {
-				Type:        schema.TypeString,
+				ForceNew:    true,
 				Computed:    true,
-				Description: "Log theme ID.",
+				Description: "The name of the log topic is required when `topic_id` is not filled in. If the log theme does not exist, it will be automatically created.",
 			},
 
 			"extend": {
@@ -86,49 +81,65 @@ func ResourceTencentCloudClsCloudProductLogTask() *schema.Resource {
 				Computed:    true,
 				Description: "Log configuration extension information, generally used to store additional log delivery configurations.",
 			},
+
+			"logset_id": {
+				Type:        schema.TypeString,
+				Optional:    true,
+				ForceNew:    true,
+				Computed:    true,
+				Description: "Log set ID.",
+			},
+
+			"topic_id": {
+				Type:        schema.TypeString,
+				Optional:    true,
+				ForceNew:    true,
+				Computed:    true,
+				Description: "Log theme ID.",
+			},
+
+			"force_delete": {
+				Type:        schema.TypeBool,
+				Optional:    true,
+				Description: "Indicate whether to forcibly delete the corresponding logset and topic. If set to true, it will be forcibly deleted. Default is false.",
+			},
 		},
 	}
 }
 
-func resourceTencentCloudClsCloudProductLogTaskCreate(d *schema.ResourceData, meta interface{}) error {
-	defer tccommon.LogElapsed("resource.tencentcloud_cls_cloud_product_log_task.create")()
+func resourceTencentCloudClsCloudProductLogTaskV2Create(d *schema.ResourceData, meta interface{}) error {
+	defer tccommon.LogElapsed("resource.tencentcloud_cls_cloud_product_log_task_v2.create")()
 	defer tccommon.InconsistentCheck(d, meta)()
 
-	logId := tccommon.GetLogId(tccommon.ContextNil)
-
-	ctx := tccommon.NewResourceLifeCycleHandleFuncContext(context.Background(), logId, d, meta)
-
 	var (
+		logId              = tccommon.GetLogId(tccommon.ContextNil)
+		ctx                = tccommon.NewResourceLifeCycleHandleFuncContext(context.Background(), logId, d, meta)
+		request            = clsv20201016.NewCreateCloudProductLogCollectionRequest()
 		instanceId         string
 		assumerName        string
 		logType            string
 		cloudProductRegion string
 	)
-	var (
-		request  = clsv20201016.NewCreateCloudProductLogCollectionRequest()
-		response = clsv20201016.NewCreateCloudProductLogCollectionResponse()
-	)
 
 	if v, ok := d.GetOk("instance_id"); ok {
+		request.InstanceId = helper.String(v.(string))
 		instanceId = v.(string)
 	}
+
 	if v, ok := d.GetOk("assumer_name"); ok {
+		request.AssumerName = helper.String(v.(string))
 		assumerName = v.(string)
 	}
+
 	if v, ok := d.GetOk("log_type"); ok {
+		request.LogType = helper.String(v.(string))
 		logType = v.(string)
 	}
+
 	if v, ok := d.GetOk("cloud_product_region"); ok {
+		request.CloudProductRegion = helper.String(v.(string))
 		cloudProductRegion = v.(string)
 	}
-
-	request.InstanceId = helper.String(instanceId)
-
-	request.AssumerName = helper.String(assumerName)
-
-	request.LogType = helper.String(logType)
-
-	request.CloudProductRegion = helper.String(cloudProductRegion)
 
 	if v, ok := d.GetOk("cls_region"); ok {
 		request.ClsRegion = helper.String(v.(string))
@@ -146,6 +157,14 @@ func resourceTencentCloudClsCloudProductLogTaskCreate(d *schema.ResourceData, me
 		request.Extend = helper.String(v.(string))
 	}
 
+	if v, ok := d.GetOk("logset_id"); ok {
+		request.LogsetId = helper.String(v.(string))
+	}
+
+	if v, ok := d.GetOk("topic_id"); ok {
+		request.TopicId = helper.String(v.(string))
+	}
+
 	err := resource.Retry(tccommon.WriteRetryTimeout, func() *resource.RetryError {
 		result, e := meta.(tccommon.ProviderMeta).GetAPIV3Conn().UseClsV20201016Client().CreateCloudProductLogCollectionWithContext(ctx, request)
 		if e != nil {
@@ -153,97 +172,130 @@ func resourceTencentCloudClsCloudProductLogTaskCreate(d *schema.ResourceData, me
 		} else {
 			log.Printf("[DEBUG]%s api[%s] success, request body [%s], response body [%s]\n", logId, request.GetAction(), request.ToJsonString(), result.ToJsonString())
 		}
-		response = result
+
+		if result == nil || result.Response == nil {
+			return resource.NonRetryableError(fmt.Errorf("Create cls cloud product log task failed, Response is nil."))
+		}
+
 		return nil
 	})
+
 	if err != nil {
 		log.Printf("[CRITAL]%s create cls cloud product log task failed, reason:%+v", logId, err)
 		return err
 	}
 
-	_ = response
-
+	// wait
 	service := ClsService{client: meta.(tccommon.ProviderMeta).GetAPIV3Conn()}
 	conf := tccommon.BuildStateChangeConf([]string{}, []string{"1"}, 10*tccommon.ReadRetryTimeout, time.Second, service.ClsCloudProductLogTaskStateRefreshFunc(ctx, instanceId, assumerName, logType, []string{}))
-
 	if _, e := conf.WaitForState(); e != nil {
 		return e
 	}
 
 	d.SetId(strings.Join([]string{instanceId, assumerName, logType, cloudProductRegion}, tccommon.FILED_SP))
 
-	return resourceTencentCloudClsCloudProductLogTaskRead(d, meta)
+	return resourceTencentCloudClsCloudProductLogTaskV2Read(d, meta)
 }
 
-func resourceTencentCloudClsCloudProductLogTaskRead(d *schema.ResourceData, meta interface{}) error {
-	defer tccommon.LogElapsed("resource.tencentcloud_cls_cloud_product_log_task.read")()
+func resourceTencentCloudClsCloudProductLogTaskV2Read(d *schema.ResourceData, meta interface{}) error {
+	defer tccommon.LogElapsed("resource.tencentcloud_cls_cloud_product_log_task_v2.read")()
 	defer tccommon.InconsistentCheck(d, meta)()
 
-	logId := tccommon.GetLogId(tccommon.ContextNil)
-
-	ctx := tccommon.NewResourceLifeCycleHandleFuncContext(context.Background(), logId, d, meta)
-
-	service := ClsService{client: meta.(tccommon.ProviderMeta).GetAPIV3Conn()}
+	var (
+		logId       = tccommon.GetLogId(tccommon.ContextNil)
+		ctx         = tccommon.NewResourceLifeCycleHandleFuncContext(context.Background(), logId, d, meta)
+		service     = ClsService{client: meta.(tccommon.ProviderMeta).GetAPIV3Conn()}
+		deleteForce bool
+	)
 
 	idSplit := strings.Split(d.Id(), tccommon.FILED_SP)
 	if len(idSplit) != 4 {
 		return fmt.Errorf("id is broken,%s", d.Id())
 	}
+
 	instanceId := idSplit[0]
 	assumerName := idSplit[1]
 	logType := idSplit[2]
 	cloudProductRegion := idSplit[3]
-
-	_ = d.Set("instance_id", instanceId)
-
-	_ = d.Set("assumer_name", assumerName)
-
-	_ = d.Set("log_type", logType)
-
-	_ = d.Set("cloud_product_region", cloudProductRegion)
 
 	respData, err := service.DescribeClsCloudProductLogTaskById(ctx, instanceId, assumerName, logType)
 	if err != nil {
 		return err
 	}
 
-	if respData == nil {
+	if respData == nil || len(respData.Tasks) < 1 {
 		d.SetId("")
 		log.Printf("[WARN]%s resource `cls_cloud_product_log_task` [%s] not found, please check if it has been deleted.\n", logId, d.Id())
 		return nil
 	}
 
-	if len(respData.Tasks) > 0 {
-		_ = d.Set("topic_id", respData.Tasks[0].TopicId)
-		_ = d.Set("logset_id", respData.Tasks[0].LogsetId)
-		_ = d.Set("extend", respData.Tasks[0].Extend)
+	_ = d.Set("instance_id", instanceId)
+	_ = d.Set("assumer_name", assumerName)
+	_ = d.Set("log_type", logType)
+	_ = d.Set("cloud_product_region", cloudProductRegion)
+
+	if respData.Tasks[0].ClsRegion != nil {
 		_ = d.Set("cls_region", respData.Tasks[0].ClsRegion)
 	}
-	_ = instanceId
-	_ = assumerName
-	_ = logType
-	_ = cloudProductRegion
+
+	if respData.Tasks[0].Extend != nil {
+		_ = d.Set("extend", respData.Tasks[0].Extend)
+	}
+
+	if respData.Tasks[0].LogsetId != nil {
+		_ = d.Set("logset_id", respData.Tasks[0].LogsetId)
+		info, err := service.DescribeClsLogset(ctx, *respData.Tasks[0].LogsetId)
+		if err != nil {
+			return err
+		}
+
+		if info.LogsetName != nil {
+			_ = d.Set("logset_name", info.LogsetName)
+		}
+	}
+
+	if respData.Tasks[0].TopicId != nil {
+		_ = d.Set("topic_id", respData.Tasks[0].TopicId)
+		info, err := service.DescribeClsTopicById(ctx, *respData.Tasks[0].TopicId)
+		if err != nil {
+			return err
+		}
+
+		if info.TopicName != nil {
+			_ = d.Set("topic_name", info.TopicName)
+		}
+	}
+
+	if v, ok := d.GetOkExists("force_delete"); ok {
+		deleteForce = v.(bool)
+	}
+
+	_ = d.Set("force_delete", deleteForce)
+
 	return nil
 }
 
-func resourceTencentCloudClsCloudProductLogTaskUpdate(d *schema.ResourceData, meta interface{}) error {
-	defer tccommon.LogElapsed("resource.tencentcloud_cls_cloud_product_log_task.update")()
+func resourceTencentCloudClsCloudProductLogTaskV2Update(d *schema.ResourceData, meta interface{}) error {
+	defer tccommon.LogElapsed("resource.tencentcloud_cls_cloud_product_log_task_v2.update")()
 	defer tccommon.InconsistentCheck(d, meta)()
 
-	logId := tccommon.GetLogId(tccommon.ContextNil)
+	var (
+		logId = tccommon.GetLogId(tccommon.ContextNil)
+		ctx   = tccommon.NewResourceLifeCycleHandleFuncContext(context.Background(), logId, d, meta)
+	)
 
-	ctx := tccommon.NewResourceLifeCycleHandleFuncContext(context.Background(), logId, d, meta)
-
-	immutableArgs := []string{"instance_id", "assumer_name", "log_type", "cloud_product_region", "cls_region", "logset_name", "topic_name"}
+	immutableArgs := []string{"cls_region"}
 	for _, v := range immutableArgs {
 		if d.HasChange(v) {
 			return fmt.Errorf("argument `%s` cannot be changed", v)
 		}
 	}
+
 	idSplit := strings.Split(d.Id(), tccommon.FILED_SP)
 	if len(idSplit) != 4 {
 		return fmt.Errorf("id is broken,%s", d.Id())
 	}
+
 	instanceId := idSplit[0]
 	assumerName := idSplit[1]
 	logType := idSplit[2]
@@ -260,15 +312,10 @@ func resourceTencentCloudClsCloudProductLogTaskUpdate(d *schema.ResourceData, me
 
 	if needChange {
 		request := clsv20201016.NewModifyCloudProductLogCollectionRequest()
-
 		request.InstanceId = helper.String(instanceId)
-
 		request.AssumerName = helper.String(assumerName)
-
 		request.LogType = helper.String(logType)
-
 		request.CloudProductRegion = helper.String(cloudProductRegion)
-
 		if v, ok := d.GetOk("extend"); ok {
 			request.Extend = helper.String(v.(string))
 		}
@@ -280,46 +327,44 @@ func resourceTencentCloudClsCloudProductLogTaskUpdate(d *schema.ResourceData, me
 			} else {
 				log.Printf("[DEBUG]%s api[%s] success, request body [%s], response body [%s]\n", logId, request.GetAction(), request.ToJsonString(), result.ToJsonString())
 			}
+
 			return nil
 		})
+
 		if err != nil {
 			log.Printf("[CRITAL]%s update cls cloud product log task failed, reason:%+v", logId, err)
 			return err
 		}
 	}
 
-	return resourceTencentCloudClsCloudProductLogTaskRead(d, meta)
+	return resourceTencentCloudClsCloudProductLogTaskV2Read(d, meta)
 }
 
-func resourceTencentCloudClsCloudProductLogTaskDelete(d *schema.ResourceData, meta interface{}) error {
-	defer tccommon.LogElapsed("resource.tencentcloud_cls_cloud_product_log_task.delete")()
+func resourceTencentCloudClsCloudProductLogTaskV2Delete(d *schema.ResourceData, meta interface{}) error {
+	defer tccommon.LogElapsed("resource.tencentcloud_cls_cloud_product_log_task_v2.delete")()
 	defer tccommon.InconsistentCheck(d, meta)()
 
-	logId := tccommon.GetLogId(tccommon.ContextNil)
-	ctx := tccommon.NewResourceLifeCycleHandleFuncContext(context.Background(), logId, d, meta)
+	var (
+		logId       = tccommon.GetLogId(tccommon.ContextNil)
+		ctx         = tccommon.NewResourceLifeCycleHandleFuncContext(context.Background(), logId, d, meta)
+		request     = clsv20201016.NewDeleteCloudProductLogCollectionRequest()
+		deleteForce bool
+	)
 
 	idSplit := strings.Split(d.Id(), tccommon.FILED_SP)
 	if len(idSplit) != 4 {
 		return fmt.Errorf("id is broken,%s", d.Id())
 	}
+
 	instanceId := idSplit[0]
 	assumerName := idSplit[1]
 	logType := idSplit[2]
 	cloudProductRegion := idSplit[3]
 
-	var (
-		request  = clsv20201016.NewDeleteCloudProductLogCollectionRequest()
-		response = clsv20201016.NewDeleteCloudProductLogCollectionResponse()
-	)
-
 	request.InstanceId = helper.String(instanceId)
-
 	request.AssumerName = helper.String(assumerName)
-
 	request.LogType = helper.String(logType)
-
 	request.CloudProductRegion = helper.String(cloudProductRegion)
-
 	err := resource.Retry(tccommon.WriteRetryTimeout, func() *resource.RetryError {
 		result, e := meta.(tccommon.ProviderMeta).GetAPIV3Conn().UseClsV20201016Client().DeleteCloudProductLogCollectionWithContext(ctx, request)
 		if e != nil {
@@ -327,71 +372,72 @@ func resourceTencentCloudClsCloudProductLogTaskDelete(d *schema.ResourceData, me
 		} else {
 			log.Printf("[DEBUG]%s api[%s] success, request body [%s], response body [%s]\n", logId, request.GetAction(), request.ToJsonString(), result.ToJsonString())
 		}
-		response = result
+
 		return nil
 	})
+
 	if err != nil {
 		log.Printf("[CRITAL]%s delete cls cloud product log task failed, reason:%+v", logId, err)
 		return err
 	}
 
-	_ = response
-
+	// wait delete
 	service := ClsService{client: meta.(tccommon.ProviderMeta).GetAPIV3Conn()}
 	conf := tccommon.BuildStateChangeConf([]string{}, []string{"3"}, 10*tccommon.ReadRetryTimeout, time.Second, service.ClsCloudProductLogTaskStateRefreshFunc(ctx, instanceId, assumerName, logType, []string{}))
-
 	if _, e := conf.WaitForState(); e != nil {
 		return e
 	}
-	var (
-		request1  = clsv20201016.NewDeleteTopicRequest()
-		response1 = clsv20201016.NewDeleteTopicResponse()
-	)
 
-	if v, ok := d.GetOk("topic_id"); ok {
-		request1.TopicId = helper.String(v.(string))
+	if v, ok := d.GetOkExists("force_delete"); ok {
+		deleteForce = v.(bool)
 	}
 
-	err = resource.Retry(tccommon.WriteRetryTimeout, func() *resource.RetryError {
-		result, e := meta.(tccommon.ProviderMeta).GetAPIV3Conn().UseClsV20201016Client().DeleteTopicWithContext(ctx, request1)
-		if e != nil {
-			return tccommon.RetryError(e)
-		} else {
-			log.Printf("[DEBUG]%s api[%s] success, request body [%s], response body [%s]\n", logId, request1.GetAction(), request1.ToJsonString(), result.ToJsonString())
+	if deleteForce {
+		var (
+			request1 = clsv20201016.NewDeleteTopicRequest()
+			request2 = clsv20201016.NewDeleteLogsetRequest()
+		)
+
+		if v, ok := d.GetOk("topic_id"); ok {
+			request1.TopicId = helper.String(v.(string))
 		}
-		response1 = result
-		return nil
-	})
-	if err != nil {
-		log.Printf("[CRITAL]%s delete cls cloud product log task failed, reason:%+v", logId, err)
-		return err
-	}
 
-	_ = response1
-	var (
-		request2  = clsv20201016.NewDeleteLogsetRequest()
-		response2 = clsv20201016.NewDeleteLogsetResponse()
-	)
+		err = resource.Retry(tccommon.WriteRetryTimeout, func() *resource.RetryError {
+			result, e := meta.(tccommon.ProviderMeta).GetAPIV3Conn().UseClsV20201016Client().DeleteTopicWithContext(ctx, request1)
+			if e != nil {
+				return tccommon.RetryError(e)
+			} else {
+				log.Printf("[DEBUG]%s api[%s] success, request body [%s], response body [%s]\n", logId, request1.GetAction(), request1.ToJsonString(), result.ToJsonString())
+			}
 
-	if v, ok := d.GetOk("logset_id"); ok {
-		request2.LogsetId = helper.String(v.(string))
-	}
+			return nil
+		})
 
-	err = resource.Retry(tccommon.WriteRetryTimeout, func() *resource.RetryError {
-		result, e := meta.(tccommon.ProviderMeta).GetAPIV3Conn().UseClsV20201016Client().DeleteLogsetWithContext(ctx, request2)
-		if e != nil {
-			return tccommon.RetryError(e)
-		} else {
-			log.Printf("[DEBUG]%s api[%s] success, request body [%s], response body [%s]\n", logId, request2.GetAction(), request2.ToJsonString(), result.ToJsonString())
+		if err != nil {
+			log.Printf("[CRITAL]%s delete cls cloud product log task topic failed, reason:%+v", logId, err)
+			return err
 		}
-		response2 = result
-		return nil
-	})
-	if err != nil {
-		log.Printf("[CRITAL]%s delete cls cloud product log task failed, reason:%+v", logId, err)
-		return err
+
+		if v, ok := d.GetOk("logset_id"); ok {
+			request2.LogsetId = helper.String(v.(string))
+		}
+
+		err = resource.Retry(tccommon.WriteRetryTimeout, func() *resource.RetryError {
+			result, e := meta.(tccommon.ProviderMeta).GetAPIV3Conn().UseClsV20201016Client().DeleteLogsetWithContext(ctx, request2)
+			if e != nil {
+				return tccommon.RetryError(e)
+			} else {
+				log.Printf("[DEBUG]%s api[%s] success, request body [%s], response body [%s]\n", logId, request2.GetAction(), request2.ToJsonString(), result.ToJsonString())
+			}
+
+			return nil
+		})
+
+		if err != nil {
+			log.Printf("[CRITAL]%s delete cls cloud product log task logset failed, reason:%+v", logId, err)
+			return err
+		}
 	}
 
-	_ = response2
 	return nil
 }
